# apps/server/swagger.json
{
  "openapi": "3.0.0",
  "components": {
...
    "schemas": {
      "ApiResponse_null_": {
        "description": "Standard API Response structure.\nThis class creates a consistent envelope for all API responses.",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "anyOf": [],
            "nullable": true
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": ["success", "message", "data", "timestamp"],
        "type": "object",
        "additionalProperties": false
      },
...
      "UserDTO": {
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "allowRemote": {
            "type": "boolean"
          },
          "type": {
            "type": "string"
          },
          "allowVideoTranscoding": {
            "type": "boolean"
          },
          "internetBitrateLimit": {
            "type": "number",
            "format": "double"
          },
          "allowDownloads": {
            "type": "boolean"
          },
          "hideInLogin": {
            "type": "boolean"
          },
          "maxSessions": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "id",
          "username",
          "allowRemote",
          "type",
          "allowVideoTranscoding",
          "allowDownloads",
          "hideInLogin",
          "maxSessions"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "ApiResponse__token-string--user-UserDTO__": {
        "description": "Standard API Response structure.\nThis class creates a consistent envelope for all API responses.",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "properties": {
              "user": {
                "$ref": "#/components/schemas/UserDTO"
              },
              "token": {
                "type": "string"
              }
            },
            "required": ["user", "token"],
            "type": "object",
            "nullable": true
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": ["success", "message", "data", "timestamp"],
        "type": "object",
        "additionalProperties": false
      },
      "UserType": {
        "enum": ["normal", "admin"],
        "type": "string"
      },
      "CreateUserDTO": {
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "allowRemote": {
            "type": "boolean"
          },
          "type": {
            "$ref": "#/components/schemas/UserType"
          },
          "allowVideoTranscoding": {
            "type": "boolean"
          },
          "internetBitrateLimit": {
            "type": "number",
            "format": "double"
          },
          "allowDownloads": {
            "type": "boolean"
          },
          "hideInLogin": {
            "type": "boolean"
          },
          "maxSessions": {
            "type": "number",
            "format": "double"
          }
        },
        "required": ["username"],
        "type": "object",
        "additionalProperties": false
      },
      "ApiResponse_UserDTO_": {
        "description": "Standard API Response structure.\nThis class creates a consistent envelope for all API responses.",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserDTO"
              }
            ],
            "nullable": true
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": ["success", "message", "data", "timestamp"],
        "type": "object",
        "additionalProperties": false
      },
      "UpdateUserDTO": {
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "allowRemote": {
            "type": "boolean"
          },
          "type": {
            "$ref": "#/components/schemas/UserType"
          },
          "allowVideoTranscoding": {
            "type": "boolean"
          },
          "internetBitrateLimit": {
            "type": "number",
            "format": "double"
          },
          "allowDownloads": {
            "type": "boolean"
          },
          "hideInLogin": {
            "type": "boolean"
          },
          "maxSessions": {
            "type": "number",
            "format": "double"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "ApiResponse_UserDTO-Array_": {
        "description": "Standard API Response structure.\nThis class creates a consistent envelope for all API responses.",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/UserDTO"
            },
            "type": "array",
            "nullable": true
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": ["success", "message", "data", "timestamp"],
        "type": "object",
        "additionalProperties": false
      },
      "LoginResponseDTO": {
        "properties": {
          "user": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserDTO"
              }
            ],
            "nullable": true
          },
          "token": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        },
        "required": ["user"],
        "type": "object",
        "additionalProperties": false
      },
      "ApiResponse_LoginResponseDTO_": {
        "description": "Standard API Response structure.\nThis class creates a consistent envelope for all API responses.",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/LoginResponseDTO"
              }
            ],
            "nullable": true
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": ["success", "message", "data", "timestamp"],
        "type": "object",
        "additionalProperties": false
      },
      "LoginDTO": {
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": ["username", "password"],
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt",
        "description": "Full user authentication with library and remote access checks"
      },
      "cookieAuthFast": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt",
        "description": "Fast user authentication without additional checks"
      },
      "adminAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt",
        "description": "Admin user authentication"
      },
      "managementAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt",
        "description": "Management access for local or admin users"
      },
      "public": {
        "type": "apiKey",
        "in": "header",
        "name": "x-public-access",
        "description": "Public access without authentication"
      }
    }
  },
  "info": {
    "title": "Seerial Media Server API",
    "version": "0.2.0",
    "description": "Media management server API for the Seerial suite",
    "license": {
      "name": "GPL-3.0-only"
    },
    "contact": {
      "name": "Juan Llavero"
    }
  },
  "paths": {
...
    "/users": {
      "post": {
        "operationId": "Create",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse__token-string--user-UserDTO__"
                }
              }
            }
          }
        },
        "description": "Create a new user",
        "tags": ["Users"],
        "security": [
          {
            "managementAuth": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDTO"
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "patch": {
        "operationId": "Update",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_UserDTO_"
                }
              }
            }
          }
        },
        "description": "Update user details",
        "tags": ["Users"],
        "security": [
          {
            "managementAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDTO"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "Delete",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_null_"
                }
              }
            }
          }
        },
        "description": "Delete a user",
        "tags": ["Users"],
        "security": [
          {
            "managementAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/users/public": {
      "get": {
        "operationId": "FindAll",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_UserDTO-Array_"
                }
              }
            }
          }
        },
        "description": "Get all users (public access)",
        "tags": ["Users"],
        "security": [
          {
            "public": []
          }
        ],
        "parameters": []
      }
    },
    "/users/login": {
      "post": {
        "operationId": "Login",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_LoginResponseDTO_"
                }
              }
            }
          }
        },
        "description": "User login",
        "tags": ["Users"],
        "security": [
          {
            "public": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            }
          }
        }
      }
    },
    "/users/logout": {
      "post": {
        "operationId": "Logout",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_null_"
                }
              }
            }
          }
        },
        "description": "User logout",
        "tags": ["Users"],
        "security": [
          {
            "public": []
          }
        ],
        "parameters": []
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:34200/api"
    }
  ]
}